move: 메인 페이지를 index.tsx로 선언하기
